+++ TinkoffASDKCore/TinkoffASDKCore/IP/IPAddressProvider.swift
@@ -1,0 +1,71 @@
+//
+//
+//  IPAddressProvider.swift
+//
+//  Copyright (c) 2021 Tinkoff Bank
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+
+import Foundation
+
+public struct IPAddressProvider {
+    var ipAddress: IPAddress? {
+        return getDeviceIPAddresses().first
+    }
+    
+    private let factory: IPAddressFactory
+    
+    init(factory: IPAddressFactory) {
+        self.factory = factory
+    }
+}
+
+private extension IPAddressProvider {
+    func getDeviceIPAddresses() -> [IPAddress] {
+        let addressStrings = getDeviceIPAddressStrings()
+        return addressStrings.compactMap { factory.ipAddress(with: $0) }
+    }
+    
+    func getDeviceIPAddressStrings() -> [String] {
+        var addresses: [String] = []
+
+        // Get list of all interfaces on the local machine:
+        var ifaddr: UnsafeMutablePointer<ifaddrs>?
+        guard getifaddrs(&ifaddr) == 0 else { return [] }
+        guard let firstAddr = ifaddr else { return [] }
+
+        // For each interface ...
+        for ptr in sequence(first: firstAddr, next: { $0.pointee.ifa_next }) {
+            let flags = Int32(ptr.pointee.ifa_flags)
+            let addr = ptr.pointee.ifa_addr.pointee
+
+            // Check for running IPv4, IPv6 interfaces. Skip the loopback interface.
+            if (flags & (IFF_UP | IFF_RUNNING | IFF_LOOPBACK)) == (IFF_UP | IFF_RUNNING) {
+                if addr.sa_family == UInt8(AF_INET) || addr.sa_family == UInt8(AF_INET6) {
+                    // Convert interface address to a human readable string:
+                    var hostname = [CChar](repeating: 0, count: Int(NI_MAXHOST))
+                    if getnameinfo(ptr.pointee.ifa_addr, socklen_t(addr.sa_len), &hostname, socklen_t(hostname.count), nil, socklen_t(0), NI_NUMERICHOST) == 0 {
+                        let address = String(cString: hostname)
+                        addresses.append(address)
+                    }
+                }
+            }
+        }
+
+        freeifaddrs(ifaddr)
+
+        return addresses
+    }
+}
