+++ TinkoffASDKUI/TinkoffASDKUI/AcquiringUISDK.swift
@@ -949,7 +949,7 @@ public class AcquiringUISDK: NSObject {
                                                       screenHeight: Int(screenSize.height))
                     finistRequestData.setDeviceInfo(info: deviceInfo)
                     finistRequestData.setThreeDSVersion(checkResult.version)
-                    finistRequestData.setIpAddress(self.acquiringSdk.networkIpAddress())
+                    finistRequestData.setIpAddress(self.acquiringSdk.networkIpAddress()?.fullStringValue)
                 }
                 // завершаем оплату
                 self.finishAuthorize(requestData: finistRequestData, treeDSmessageVersion: checkResult.version) { finishResponse in
